requirements.txt
flask==2.2.5
werkzeug==2.2.3
openai-whisper
torch
googletrans==4.0.0-rc1